Added additional logic to admin page and clerk authentication

+++ src/services/photo.service.ts
@@ -1,11 +1,12 @@
+import { getToken } from "@clerk/react";
 import type { BaseResponse } from "../types/BaseResponse";
 import type { Image } from "../types/Image";
+import type { UploadImagePayload } from "../types/UploadImagePayload";
 
 const BASE_URL = `${import.meta.env.VITE_API_BASE_URL}/api/v1`;
 
 export async function getAllPhotos() {
   const res: Response = await fetch(`${BASE_URL}/photo`);
-  console.log(res);
 
   if (!res.ok) {
     throw new Error("Failed to fetch images");
@@ -15,16 +16,13 @@ export async function getAllPhotos() {
   return json.data;
 }
 
-interface UploadImagePayload {
-  imageBase64: string;
-  caption: string;
-  metaData: string;
-}
-
 export async function uploadImage(payload: UploadImagePayload) {
+  const sessionToken = await getToken();
+  if (!sessionToken) throw new Error("Unauthorized");
+
   const res: Response = await fetch(`${BASE_URL}/photo`, {
     method: "POST",
-    headers: { "Content-Type": "application/json" },
+    headers: { "Content-Type": "application/json", Authorization: `Bearer ${sessionToken}` },
     body: JSON.stringify(payload),
   });
 
@@ -37,8 +35,12 @@ export async function uploadImage(payload: UploadImagePayload) {
 }
 
 export async function deleteImage(imageId: string) {
+  const sessionToken = await getToken();
+  if (!sessionToken) throw new Error("Unauthorized");
+
   const res: Response = await fetch(`${BASE_URL}/photo/${imageId}`, {
     method: "DELETE",
+    headers: { Authorization: `Bearer ${sessionToken}` },
   });
 
   if (!res.ok) {
